fix: resolve MSVC C4244/C4996/C5105 build diagnostics for VS2019
Agent-Logs-Url: https://github.com/Zero3K20/steam-guard/sessions/5dc9c696-c62a-4b03-8930-f54111a18fc0

+++ src/server_time.c
@@ -92,7 +92,7 @@ static int get_server_time_diff_winhttp()
     if (0 == server_time)
         return 0;
 
-    return server_time - time(NULL);
+    return (int)(server_time - time(NULL));
 }
 #else
 static size_t write_cb(void *data, size_t size, size_t nmemb, void *clientp)
@@ -167,5 +167,5 @@ int get_server_time_diff()
 
 int get_steam_time(int server_time_diff)
 {
-    return time(NULL) + server_time_diff;
+    return (int)(time(NULL) + server_time_diff);
 } // get_steam_time
